Renamed ros-interface into deburring_ros_interface
