+++ deburring-ros-interface/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.1)
 
 SET(PROJECT_NAMESPACE ComePerrot)
-SET(PROJECT_NAME ros-interface)
+SET(PROJECT_NAME deburring_ros_interface)
 SET(PROJECT_DESCRIPTION "Ros interface for Talos manipulation MPC")
 SET(PROJECT_URL https://github.com/${PROJECT_NAMESPACE}/talos-manipulation)
