+++ deburring-ros-interface/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>ros-interface</name>
+  <name>deburring_ros_interface</name>
   <version>0.0.0</version>
   <description>Ros interface for Talos manipulation MPC</description>
