--- ros-interface/launch/ros-mpc-basic.launch
@@ -1,4 +1,0 @@
-<launch>
-    <rosparam command="load" file="$(find ros-interface)/config/ros-mpc-pointing.yaml"/>
-    <node pkg="ros-interface" type="ros_mpc_pointing" name="ros_mpc_pointing" output="screen"/>
-</launch>
